Conflicts:
	jfreereport/java/flute/makefile.mk
	jfreereport/java/libbase/makefile.mk
	jfreereport/java/libfonts/makefile.mk
	jfreereport/java/libformula/makefile.mk
	jfreereport/java/libloader/makefile.mk
	jfreereport/java/librepository/makefile.mk
	jfreereport/java/libserializer/makefile.mk
	jfreereport/java/libxml/makefile.mk
	jfreereport/java/sac/makefile.mk
	mysqlc/source/mysqlc_connection.cxx
	mysqlc/source/mysqlc_connection.hxx
	mysqlc/source/mysqlc_databasemetadata.cxx
	mysqlc/source/mysqlc_databasemetadata.hxx
	mysqlc/source/mysqlc_driver.cxx
	mysqlc/source/mysqlc_driver.hxx
	mysqlc/source/mysqlc_general.hxx
	mysqlc/source/mysqlc_preparedstatement.hxx
	mysqlc/source/mysqlc_resultsetmetadata.hxx
	mysqlc/source/mysqlc_statement.hxx
	sdext/source/minimizer/impoptimizer.cxx
	sdext/source/pdfimport/config/description.xml
	sdext/source/pdfimport/tree/drawtreevisiting.cxx
	sdext/source/presenter/PresenterWindowManager.cxx
	swext/prj/build.lst

+++ sdext/source/pdfimport/pdfparse/pdfparse.cxx
@@ -576,7 +576,7 @@ PDFEntry* PDFReader::read( const char* pBuffer, unsigned int nLen )
                  aInfo.stop, pBuffer, aInfo.stop - pBuffer,
                  aInfo.hit ? "true" : "false",
                  aInfo.full ? "true" : "false",
-                 aInfo.length );
+                 (int)aInfo.length );
         #endif
     }
     catch( parser_error<const char*, const char*>& rError )
@@ -656,7 +656,7 @@ PDFEntry* PDFReader::read( const char* pFileName )
                  aInfo.stop - file_start,
                  aInfo.hit ? "true" : "false",
                  aInfo.full ? "true" : "false",
-                 aInfo.length );
+                 (int)aInfo.length );
         #endif
     }
     catch( parser_error< const char*, file_iterator<> >& rError )
